Cover the cases the specification does not, and keep the guardrails
The twenty specified behaviours are a floor. What they leave untested is
mostly the ways this module fails rather than the ways it works: two
syncs racing, data changing mid-import, a purchase and its refund that
look exactly like the two halves of a transfer, a ciphertext lifted into
another user's row, a category value nobody has ever seen.

Two things were found by writing them. Spring logs the outgoing request
object at debug level, and a record prints every component, so an access
token was being written to the log in full whenever debug logging was
on; credentials now travel wrapped in a type whose toString says
nothing. And the architecture rule against raw decimals in the domain
caught a BigDecimal constant of mine, which is the rule working rather
than the rule being in the way.

The store is held to one contract run against both the in-memory
implementation the fast tests use and real Postgres, so the properties
those tests lean on are shown rather than assumed.

+++ budget-core/src/main/java/com/hasan/budget/ingestion/domain/Frequency.java
@@ -20,7 +20,13 @@ public enum Frequency {
     /** The provider saw a repeat it could not put a cadence on. Treated as monthly. */
     UNKNOWN(12);
 
-    private static final BigDecimal MONTHS_PER_YEAR = BigDecimal.valueOf(12);
+    /**
+     * Held as an int rather than a {@code BigDecimal} constant on purpose: an architecture rule bans
+     * raw decimal fields anywhere in the domain, because that is how amounts escape {@link Money}
+     * and take the scale and rounding bugs it exists to prevent with them. The conversion below
+     * needs a decimal for one division, and a local is where that belongs.
+     */
+    private static final int MONTHS_PER_YEAR = 12;
 
     private final int timesPerYear;
 
@@ -44,7 +50,7 @@ public enum Frequency {
         BigDecimal monthly = perOccurrence
                 .amount()
                 .multiply(BigDecimal.valueOf(timesPerYear))
-                .divide(MONTHS_PER_YEAR, 2, RoundingMode.HALF_UP);
+                .divide(BigDecimal.valueOf(MONTHS_PER_YEAR), 2, RoundingMode.HALF_UP);
         // Already at the scale Money keeps, so wrapping it rounds nothing a second time.
         return new Money(monthly);
     }
